+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { Global } from '@emotion/react';
 
-import { OverlayProvider } from '@ui/components/layout/overlay/OverlayProvider';
+import { OverlayProvider } from '@ui/components/overlay/OverlayProvider';
 import globalStyle from '@ui/styles/globalStyle.css';
 import routes from './routes';
