+++ src/pages/CalendarPage.tsx
@@ -1,7 +1,7 @@
-import CalendarHeader from '@ui/components/layout/calendar/CalendarHeader';
+import CalendarHeader from '@ui/components/calendar/CalendarHeader';
 import Body from '@ui/components/layout/Body';
-import DateColumn from '@ui/components/layout/calendar/DateColumn';
-import Menu from '@ui/components/layout/MenuBar/Menu';
+import DateColumn from '@ui/components/calendar/DateColumn';
+import Menu from '@ui/components/MenuBar/Menu';
 import styled from '@emotion/styled';
 import styleTokenCss from '@ui/styles/styleToken.css';
 import range from '@lib/utils/range';
