+++ src/ui/components/common/NavigationHeader.tsx
