+++ src/ui/components/diary/EmotionContainer.tsx
