+++ src/ui/components/layout/RouterLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Layout, Page } from '@ui/components/layout';
+import { Layout, Page } from '@ui/components/layout/index';
 
 export default function RouterLayout() {
   return (
